リファクタリング: 検索処理を責務別に分割し、バージョンを 0.8.4 に更新
## 概要

`search.ts` に集約されていた検索関連処理を、責務ごとのモジュールへ分割しました。

あわせて、パッケージバージョンを `0.8.3` から `0.8.4` に更新しました。

## 変更内容

- context lines 生成処理を `context-lines.ts` に分離
- エンコーディング選択と decode 処理を `encoding.ts` に分離
- テキストマッチ、スニペット生成、行分割処理を `match-text.ts` に分離
- 検索結果の集約、summary/detail matches 構築、truncation diagnostics 処理を `search-results.ts` に分離
- `search.ts` 側は traversal と file scan の制御を中心に整理
- `package.json` と `package-lock.json` のバージョンを `0.8.4` に更新

## 確認

- `npm test`
- `npm run build`
