+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "miku-grep",
-  "version": "0.8.3",
+  "version": "0.8.4",
   "description": "Local-first structured grep CLI for AI agents and automation.",
   "license": "Apache-2.0",
   "type": "module",
